FIX module to validate PO on confirmation and hides/show button
