+++ tko_purchase_supplier_min_purchase_value/__init__.py
@@ -23,3 +23,4 @@
 ##############################################################################
 
 import product
+import purchase
