+++ tko_purchase_supplier_min_purchase_value/__openerp__.py
@@ -38,6 +38,7 @@
     ],
     'data': [
         'product_view.xml',
+        'purchase_view.xml',
     ],
     'init': [],
     'demo': [],
