+++ tko_purchase_supplier_min_purchase_value/purchase_view.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<openerp>
+    <data>
+        <record id="purchase_order_form_view_inherit" model="ir.ui.view">
+            <field name="name">purchase.order.form</field>
+            <field name="model">purchase.order</field>
+            <field name="inherit_id" ref="purchase.purchase_order_form"/>
+            <field name="arch" type="xml">
+                <xpath expr="//field[@name='partner_id']" position="after">
+                    <field name="is_visible" invisible="1"/>
+                </xpath>
+                <xpath expr="//header/button[@name='purchase_confirm']" position="replace">
+                    <button name="purchase_confirm" string="Confirm Order" id="draft_confirm"
+                            attrs="{'invisible':[('is_visible','=',False),('state','=','draft')]}"/>
+                </xpath>
+            </field>
+        </record>
+
+    </data>
+</openerp>
